requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
youtube-dl==2021.12.17